% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_noFCmsgPlot.R
\name{app_noFCmsgPlot}
\alias{app_noFCmsgPlot}
\title{No Fold Change message plot}
\usage{
app_noFCmsgPlot()
}
\value{
This returns a plot obejct with a message.
}
\description{
This function is used by the app when no Fold Chnage is provided.
}
