% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_app_input_utils.R
\name{app_parse_textarea}
\alias{app_parse_textarea}
\title{Parse text area}
\usage{
app_parse_textarea(gene_string)
}
\arguments{
\item{gene_string}{This a string with only gene ids separated by new line character.
Optionally it also accepts fold change (fc) next to each gene separated by coma.}
}
\value{
This returns a list containing
  * Gene vector
  * Gene with fc. If fc not provided. this is NULL.
}
\description{
Takes user input and parse into list, which can be used by SigBio application.
}
