% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_enrichGO_utils.R
\name{do_enrichGO}
\alias{do_enrichGO}
\title{Do enrichGO}
\usage{
do_enrichGO(...)
}
\arguments{
\item{...}{Arguments pass to \code{\link[clusterProfiler]{enrichGO}} and 
\code{\link[clusterProfiler]{setReadable}}}
}
\value{
A list of enrichGO objects
}
\description{
This is a wrapper function to do \code{\link[clusterProfiler]{enrichGO}} on 
all ontology terms and convert the gene IDs to gene symbles.
}
