% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ids.R
\name{do_selectIds}
\alias{do_selectIds}
\title{select all}
\usage{
do_selectIds(genelist, org_pkg, gtf_type)
}
\arguments{
\item{genelist}{A vector with list of input genes}

\item{org_pkg}{An AnnotationHub OrgDb object}

\item{gtf_type}{Input gene type}
}
\value{
A dataframe with selected IDs
}
\description{
Wrapper funtion to \code{\link[AnnotationDbi]{select}}
}
