% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dataobj.R
\docType{data}
\name{entrez_ids_with_fc_vector}
\alias{entrez_ids_with_fc_vector}
\title{Test data from DOSE package.}
\format{A data frame
#' \describe{
  ...
}}
\usage{
entrez_ids_with_fc_vector
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
