% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_app_input_utils.R
\name{kegg_rest}
\alias{kegg_rest}
\title{Call KEGG rest API}
\usage{
kegg_rest(rest_url)
}
\arguments{
\item{rest_url}{KEGG rest url Endpoint.(https://www.kegg.jp/kegg/rest/keggapi.html)}
}
\description{
Call KEGG rest API
}
