% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapids.R
\name{mapIds_all}
\alias{mapIds_all}
\title{Wrapper funtion to AnnotationDbi::select}
\usage{
mapIds_all(genelist, org_pkg, gtf_type)
}
\arguments{
\item{genelist}{A vector with list of input genes}

\item{org_pkg}{An AnnotationHub OrgDb object}

\item{gtf_type}{Input gene type}
}
\description{
Wrapper funtion to AnnotationDbi::select
}
