% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{pathway_gse}
\alias{pathway_gse}
\title{Do gene set enrichment}
\usage{
pathway_gse(id_with_fc_list, organism = "hsa", pval = "0.05")
}
\arguments{
\item{id_with_fc_list}{A vector with gene_ids and their fold change. Much like geneList from DOSE pakage}

\item{organism}{A kegg org short name}

\item{pval}{P-value cut-off}
}
\value{
A GSE plot
}
\description{
Wraper function to \code{\link[clusterProfiler]{gseKEGG}} with some extra functionality
specific to application.
}
