% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utils.R
\name{sigbio_message}
\alias{sigbio_message}
\title{Custom message for Sig-Bio-Shiny-App}
\usage{
sigbio_message(msg, ...)
}
\arguments{
\item{msg}{The message from app}

\item{...}{Additional param to \code{\link[base]{paste}}}
}
\description{
Custom message for Sig-Bio-Shiny-App
}
