% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wego_plot.R
\name{wego_plot}
\alias{wego_plot}
\title{Plot WEGO}
\usage{
wego_plot(BP = go_table, CC = go_table, MF = go_table)
}
\arguments{
\item{BP}{An enrichGO object for Biological Process.}

\item{CC}{An enrichGO object for Cellular Components.}

\item{MF}{An enrichGO object for Molecular Functions.}
}
\value{
A WEGO alike ggplot plot object.
}
\description{
This make a WEGO alike plot.
}
